+++ src/SFA.DAS.Data.Database/Tables/DAS_ConsistencyCheck.sql
@@ -3,8 +3,8 @@
 	[Id] BIGINT NOT NULL PRIMARY KEY IDENTITY(1, 1), 
     [DataType] VARCHAR(50) NOT NULL, 
     [CheckedDateTime] DATETIME NOT NULL, 
-    [SourceSystemCount] INT NOT NULL, 
-    [RdsCount] INT NOT NULL
+    [SourceSystemCount] BIGINT NOT NULL, 
+    [RdsCount] BIGINT NOT NULL
 )
 GO
 CREATE INDEX [IX_ConsistencyCheck_DataType] ON [Data_Load].[DAS_ConsistencyCheck] ([DataType])
